+++ ComponentApi/src/main/java/com/xiaojinzi/component/anno/ServiceAnno.java
@@ -8,8 +8,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * 这是一个标识模块的生命周期类的注解,因为用注解标识的类肯定在某一个模块,所以标识了就知道那个类是哪个模块下面的
- * 所标记的类一定会在主线程上被创建. 所以可以放心在构造函数中写代码
+ * 标记一个接口
  */
 @MainThreadCreateAnno
 @Target({ElementType.TYPE, ElementType.METHOD})
@@ -18,15 +17,11 @@ public @interface ServiceAnno {
 
     /**
      * 这个服务对应的接口
-     *
-     * @return
      */
     Class[] value();
 
     /**
      * 是否是单例,默认是单例模式的
-     *
-     * @return
      */
     boolean singleTon() default true;