+++ ComponentImpl/src/main/java/com/xiaojinzi/component/support/ObjectToJsonConverter.java
@@ -1,0 +1,16 @@
+package com.xiaojinzi.component.support;
+
+import android.support.annotation.NonNull;
+
+/**
+ * Json 转化器
+ */
+public interface ObjectToJsonConverter {
+
+    /**
+     * 对象转 Json String
+     */
+    @NonNull
+    String toJson(@NonNull Object obj);
+
+}