+++ Module2/src/main/java/com/xiaojinzi/component2/service/TestInterface.kt
@@ -1,0 +1,3 @@
+package com.xiaojinzi.component2.service
+
+interface TestInterface