+++ app/src/main/java/com/xiaojinzi/componentdemo/App.java
@@ -3,6 +3,7 @@ package com.xiaojinzi.componentdemo;
 import android.app.Application;
 import androidx.annotation.NonNull;
 
+import com.google.gson.Gson;
 import com.xiaojinzi.component.Component;
 import com.xiaojinzi.component.Config;
 import com.xiaojinzi.component.impl.application.ModuleManager;
@@ -14,7 +15,6 @@ public class App extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
-
         long startTime = System.currentTimeMillis();
         // 初始化组件化相关
         Component.init(
@@ -30,18 +30,18 @@ public class App extends Application {
                         .tipWhenUseApplication(true)
                         // 开启启动优化, 必须配套使用 Gradle 插件
                         .optimizeInit(true)
-                        // 自动加载所有模块
+                        // 自动加载所有模块, 依赖上面的 optimizeInit(true)
                         .autoRegisterModule(true)
+                        // demo 测试, 线上并没有, 请勿配置
+                        .objectToJsonConverter(obj -> new Gson().toJson(obj))
+                        // 执行构建
                         .build()
         );
         long endTime = System.currentTimeMillis();
-
         LogUtil.log("---------------------------------耗时：" + (endTime - startTime));
-
         // 如果你依赖了 rx 版本, 请加上这句配置. 忽略一些不想处理的错误
         // 如果不是 rx 的版本, 请忽略
         RxErrorIgnoreUtil.ignoreError();
-
         if (BuildConfig.DEBUG) {
             ModuleManager.getInstance().check();
         }